+++ extension/js/common/extension.ts
@@ -7,10 +7,10 @@ import { DiagnoseMsgPubkeysResult, DecryptResult, VerifyRes, PgpMsgTypeResult, P
 import { FlatTypes, GlobalIndex, GlobalStore, AccountIndex, AccountStore } from './platform/store.js';
 import { Ui, Env, Browser } from './browser.js';
 import { Catch } from './platform/catch.js';
-import { AuthRes } from './api/google.js';
 import { Buf } from './core/buf.js';
 import { AjaxErr } from './api/api.js';
 import { PassphraseDialogType } from './xss_safe_factory.js';
+import { AuthRes } from './api/google-auth.js';
 
 export type GoogleAuthWindowResult$result = 'Success' | 'Denied' | 'Error' | 'Closed';
