+++ tooling/bundle-content-scripts.ts
@@ -34,6 +34,8 @@ buildContentScript(([] as string[]).concat(
   getFilesInDir(`${sourceDir}/common/platform`, /\.js$/, false),
   getFilesInDir(`${sourceDir}/common/core`, /\.js$/, false),
   getFilesInDir(`${sourceDir}/common/api`, /\.js$/, false),
+  getFilesInDir(`${sourceDir}/common/api/email_provider`, /\.js$/, false),
+  getFilesInDir(`${sourceDir}/common/api/email_provider/gmail`, /\.js$/, false),
   getFilesInDir(`${sourceDir}/common`, /\.js$/, false),
   getFilesInDir(`${sourceDir}/content_scripts/webmail`, /\.js$/),
 ), 'webmail_bundle.js');
